Update ItemController Actions to Consider User

+++ WishList/Controllers/ItemController.cs
@@ -1,21 +1,28 @@
 ﻿using System.Linq;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using WishList.Data;
+using WishList.Models;
 
 namespace WishList.Controllers
 {
+    [Authorize]
     public class ItemController : Controller
     {
+        private readonly UserManager<ApplicationUser> _userManager;
         private readonly ApplicationDbContext _context;
 
-        public ItemController(ApplicationDbContext context)
+        public ItemController(ApplicationDbContext context, UserManager<ApplicationUser> userManager)
         {
             _context = context;
+            _userManager = userManager;
         }
 
         public IActionResult Index()
         {
-            var model = _context.Items.ToList();
+            var user = _userManager.GetUserAsync(HttpContext.User).Result;
+            var model = _context.Items.Where(u=> u.User.Id == user.Id).ToList();
 
             return View("Index", model);
         }
@@ -29,17 +36,30 @@ namespace WishList.Controllers
         [HttpPost]
         public IActionResult Create(Models.Item item)
         {
+            var user = _userManager.GetUserAsync(HttpContext.User).Result;
+            item.User = user;
+
             _context.Items.Add(item);
             _context.SaveChanges();
+
             return RedirectToAction("Index");
         }
 
         public IActionResult Delete(int id)
         {
             var item = _context.Items.FirstOrDefault(e => e.Id == id);
-            _context.Items.Remove(item);
-            _context.SaveChanges();
-            return RedirectToAction("Index");
+            var user = _userManager.GetUserAsync(HttpContext.User).Result;
+
+            if (item.User == user)
+            {
+                _context.Items.Remove(item);
+                _context.SaveChanges();
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                return Unauthorized();
+            }
         }
     }
 }
